retail/recommendation-system/bqml-scann/requirements.txt
tensorflow==2.3.1
tfx==0.25.0
apache-beam[gcp]
google-cloud-bigquery 
pyarrow
google-auth 
google-api-python-client 
google-api-core
scann
kfp==1.1.2